refactor(multiverse): Move the widgets of the details mode into the details module
