+++ labs/multiverse/src/widgets/room_view/details/linked_chunk.rs
